refactor(analysis-types): split args DTO
